complete chat extension

+++ server/src/query/update.ts
@@ -47,7 +47,7 @@ export const addFriend = async(input:FriendInputProps)=>{
     
 } 
 
-export const removeFriend = async(input:FriendInputProps)=>{
+export const removeFriends = async(input:FriendInputProps)=>{
     // todo need to optimise the code if possible directly from the prisma do else we can optimise like this like getuser FUnction and fileter function and may more which are repeated more
     
     const user1 = await prisma.user.findUnique({
@@ -62,6 +62,18 @@ export const removeFriend = async(input:FriendInputProps)=>{
     })
     if(user2&& user1)
     {
+        //deleting the rooms
+        const room = await prisma.room.deleteMany({
+            where: {
+                subscribedUser: {
+                    every: {
+                        id: {
+                            in: [input.user1Id, input.user2Id]
+                        }
+                    }
+                }
+            }
+        })
         const updatedUser1Friends = user1.friends.filter((id)=>{id!==input.user2Id})
         const updatedUser2Friends = user2.friends.filter((id)=>{id!==input.user1Id})
         // removing the friends
@@ -81,24 +93,70 @@ export const removeFriend = async(input:FriendInputProps)=>{
                 friends:{set:updatedUser2Friends}
             }
         })
-        //deleting the rooms
-        const room = await prisma.room.deleteMany({
-            where:{
-                subscribedUser:{
-                    some:{
-                        id:{
-                            in:[input.user1Id,input.user2Id]
-                        }
-                    }
-                }
-            }
-        })
+        
         // console.log(updatedUser2,updatedUser1,room)
         return room
 
     }
     
 }
+export const removeFriend = async (input: FriendInputProps) => {
+    const user1 = await prisma.user.findUnique({
+        where: {
+            id: input.user1Id,
+        },
+    });
+
+    const user2 = await prisma.user.findUnique({
+        where: {
+            id: input.user2Id,
+        },
+    });
+
+    if (user1 && user2) {
+        const updatedUser1Friends = user1.friends.filter(id => id !== input.user2Id);
+        const updatedUser2Friends = user2.friends.filter(id => id !== input.user1Id);
+
+        // Update user friends
+        const updatedUser1 = await prisma.user.update({
+            where: {
+                id: input.user1Id,
+            },
+            data: {
+                friends: {
+                    set: updatedUser1Friends,
+                },
+            },
+        });
+
+        const updatedUser2 = await prisma.user.update({
+            where: {
+                id: input.user2Id,
+            },
+            data: {
+                friends: {
+                    set: updatedUser2Friends,
+                },
+            },
+        });
+
+        // Delete rooms where both users are subscribed
+        const deletedRooms = await prisma.room.deleteMany({
+            where: {
+                subscribedUser: {
+                    every: {
+                        id: {
+                            in: [input.user1Id, input.user2Id],
+                        },
+                    },
+                },
+            },
+        });
+
+        return deletedRooms;
+    }
+};
+
 export const setRequests = async(inputs:RequestProps)=>{
     const toUser = await prisma.user.findUnique({
         where:{
